requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-core==0.1.27
pydantic==2.6.1
python-dotenv==1.0.1
torch>=2.0.0
transformers>=4.36.0
sounddevice>=0.4.6
soundfile>=0.12.1
PyQt6>=6.4.0
numpy>=1.24.0
datasets>=2.14.0
accelerate>=0.26.0
ffmpeg-python>=0.2.0
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-core==0.1.27
langchain-mistralai==0.0.3
langgraph==0.0.18
boto3>=1.34.0
awscli>=1.29.0
langchain-community>=0.0.20
tavily-python>=0.3.0
replit>=3.2.4
